# packages/contracts/generated/models/GitHubCallbackSuccessResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { GitHubCallbackSuccessData } from './GitHubCallbackSuccessData';
export type GitHubCallbackSuccessResponse = {
    data: GitHubCallbackSuccessData;
};