# packages/contracts/generated/models/GitHubConnectionSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GitHubConnectionSummary = {
    id: string;
    githubLogin: string;
    githubUserId: string;
    connectedAt: string;
};