# packages/contracts/generated/models/StartGitHubConnectResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { StartGitHubConnectData } from './StartGitHubConnectData';
export type StartGitHubConnectResponse = {
    data: StartGitHubConnectData;
};