# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RepoMemory API
  version: 0.2.0
  description: Contract for currently implemented RepoMemory API endpoints.
servers:
  - url: http://localhost:8080
tags:
  - name: Health
  - name: Identity
  - name: Organizations
  - name: GitHub
paths:
  /health:
    get:
      tags:
        - Health
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: Healthy API response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/me:
    get:
      tags:
        - Identity
      operationId: getMe
      summary: Get current authenticated user
      security:
        - MockUserHeader: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /v1/organizations:
    get:
      tags:
        - Organizations
      operationId: listOrganizations
      summary: List organizations accessible by current user
      security:
        - MockUserHeader: []
      responses:
        "200":
          description: Organizations for the current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Organizations
      operationId: createOrganization
      summary: Create organization and owner membership for current user
      security:
        - MockUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/organizations/{orgId}:
    get:
      tags:
        - Organizations
      operationId: getOrganization
      summary: Get organization detail for current user
      security:
        - MockUserHeader: []
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identifier
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Organization detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/github/connect/start:
    post:
      tags:
        - GitHub
      operationId: startGitHubConnect
      summary: Start GitHub OAuth connection flow
      security:
        - MockUserHeader: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartGitHubConnectRequest"
      responses:
        "200":
          description: OAuth authorize URL ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartGitHubConnectResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "503":
          $ref: "#/components/responses/ServiceUnavailableError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/github/callback:
    get:
      tags:
        - GitHub
      operationId: completeGitHubCallback
      summary: Complete GitHub OAuth callback and persist account
      security:
        - MockUserHeader: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OAuth callback completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitHubCallbackSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "502":
          $ref: "#/components/responses/UpstreamError"
        "503":
          $ref: "#/components/responses/ServiceUnavailableError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    MockUserHeader:
      type: apiKey
      in: header
      name: x-user-id
      description: Required mock auth user identifier.
  responses:
    BadRequestError:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Missing or invalid auth headers
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenError:
      description: Authenticated user is not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UpstreamError:
      description: Upstream integration failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailableError:
      description: Service is not configured or temporarily unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
        - timestamp
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: api
        timestamp:
          type: string
          format: date-time
          example: 2026-03-06T15:04:05Z
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
          example: missing x-user-id header
        requestId:
          type: string
          example: req_12345
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    User:
      type: object
      required:
        - id
        - displayName
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
    Organization:
      type: object
      required:
        - id
        - name
        - slug
        - role
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        role:
          type: string
          enum:
            - owner
            - member
    CreateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
    StartGitHubConnectRequest:
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
    StartGitHubConnectData:
      type: object
      required:
        - redirectUrl
      properties:
        redirectUrl:
          type: string
          format: uri
    GitHubConnectionSummary:
      type: object
      required:
        - id
        - githubLogin
        - githubUserId
        - connectedAt
      properties:
        id:
          type: string
          format: uuid
        githubLogin:
          type: string
        githubUserId:
          type: string
        connectedAt:
          type: string
          format: date-time
    GitHubCallbackSuccessData:
      type: object
      required:
        - connected
        - account
      properties:
        connected:
          type: boolean
          enum: [true]
        account:
          $ref: "#/components/schemas/GitHubConnectionSummary"
    MeResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/User"
    OrganizationListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
    OrganizationResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Organization"
    StartGitHubConnectResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/StartGitHubConnectData"
    GitHubCallbackSuccessResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/GitHubCallbackSuccessData"